+++ cpp/CHANGELOG.md
@@ -111,6 +111,7 @@ Please mark all change in change log and use the ticket from JIRA.
 - MS-530 - Add unittest for SearchTask->Load
 - MS-531 - Disable next version code
 - MS-533 - Update resource_test to cover dump function
+- MS-523 - Config file validation
 
 ## New Feature
 - MS-343 - Implement ResourceMgr